+++ src/base-dto/index.ts
@@ -1,0 +1,1 @@
+export class BaseDTO {}
