+++ src/db/models/Favorite.ts
