--- src/dto/index.ts
@@ -1,3 +1,0 @@
-import { IsNumber, IsString } from 'class-validator';
-
-export class BaseDTO {}
