+++ src/modules/artist/dto/create-artist.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsBoolean, IsString } from 'class-validator';
-import { BaseDTO } from 'src/dto';
+import { BaseDTO } from 'src/base-dto';
 
 export class CreateArtistDTO extends BaseDTO {
   @IsString()
