+++ src/modules/common/common.service.ts
@@ -1,8 +1,8 @@
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { MESSAGES } from 'src/constants';
 import { DbService } from 'src/db/db.service';
-import { BaseDTO } from 'src/dto';
-import { Track } from 'src/models/Track';
+import { BaseDTO } from 'src/base-dto';
+import { Track } from 'src/db/models/Track';
 import { EDBEntryNames, ETrackRefEntry, TModelType } from 'src/types';
 
 @Injectable()
