+++ src/modules/track/dto/create-track.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsString, IsNumber, IsOptional } from 'class-validator';
-import { BaseDTO } from 'src/dto';
+import { BaseDTO } from 'src/base-dto';
 
 export class CreateTrackDTO extends BaseDTO {
   @IsString()
